update table schema

+++ src/Oasis.Application/DTOs/CabinDTO/CreateCabinRequest.cs
@@ -7,8 +7,8 @@ public sealed class CreateCabinRequest : IMapFrom<Cabin>
 {
     public string? Name { get; set; }
     public short? MaxCapacity { get; set; }
-    public short? RegularPrice { get; set; }
-    public short? Discount { get; set; }
+    public decimal? RegularPrice { get; set; }
+    public decimal? Discount { get; set; }
     public string? Description { get; set; }
     public string? Image { get; set; }
 }
