+++ src/Oasis.Domain/Cabin.cs
@@ -6,8 +6,8 @@ public class Cabin
     public DateTime CreatedAt { get; set; }
     public string? Name { get; set; }
     public short? MaxCapacity { get; set; }
-    public short? RegularPrice { get; set; }
-    public short? Discount { get; set; }
+    public decimal? RegularPrice { get; set; }
+    public decimal? Discount { get; set; }
     public string? Description { get; set; }
     public string? Image { get; set; }
 }
